rename to SelectCreatableField
