+++ tests/js/spec/components/forms/genericField.spec.jsx
@@ -4,36 +4,34 @@ import {shallow} from 'enzyme';
 import {GenericField, FormState} from 'app/components/forms';
 
 describe('GenericField', function() {
-  describe('render()', function() {
-    it('renders text as TextInput', function() {
-      let wrapper = shallow(
-        <GenericField
-          formState={FormState.READY}
-          config={{
-            name: 'field-name',
-            label: 'field label',
-            type: 'text',
-          }}
-        />
-      );
-      expect(wrapper).toMatchSnapshot();
-      expect(wrapper.name()).toEqual('TextField');
-    });
+  it('renders text as TextInput', function() {
+    let wrapper = shallow(
+      <GenericField
+        formState={FormState.READY}
+        config={{
+          name: 'field-name',
+          label: 'field label',
+          type: 'text',
+        }}
+      />
+    );
+    expect(wrapper).toMatchSnapshot();
+    expect(wrapper.name()).toEqual('TextField');
+  });
 
-    it('renders text with choices as SelectTextField', function() {
-      let wrapper = shallow(
-        <GenericField
-          formState={FormState.READY}
-          config={{
-            name: 'field-name',
-            label: 'field label',
-            type: 'text',
-            choices: [],
-          }}
-        />
-      );
-      expect(wrapper).toMatchSnapshot();
-      expect(wrapper.name()).toEqual('SelectTextField');
-    });
+  it('renders text with choices as SelectCreatableField', function() {
+    let wrapper = shallow(
+      <GenericField
+        formState={FormState.READY}
+        config={{
+          name: 'field-name',
+          label: 'field label',
+          type: 'text',
+          choices: [],
+        }}
+      />
+    );
+    expect(wrapper).toMatchSnapshot();
+    expect(wrapper.name()).toEqual('SelectCreatableField');
   });
 });
