+++ tests/js/spec/components/forms/selectCreatableField.spec.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import {mount} from 'enzyme';
 
-import {Form, SelectTextField} from 'app/components/forms';
+import {Form, SelectCreatableField} from 'app/components/forms';
 
-describe('SelectTextField', function() {
+describe('SelectCreatableField', function() {
   it('can add user input into select field when using options', function() {
     let wrapper = mount(
-      <SelectTextField options={[{value: 'foo', label: 'Foo'}]} name="fieldName" />
+      <SelectCreatableField options={[{value: 'foo', label: 'Foo'}]} name="fieldName" />
     );
 
     wrapper
@@ -32,7 +32,7 @@ describe('SelectTextField', function() {
   });
 
   it('can add user input into select field when using choices', function() {
-    let wrapper = mount(<SelectTextField choices={['foo']} name="fieldName" />);
+    let wrapper = mount(<SelectCreatableField choices={['foo']} name="fieldName" />);
 
     wrapper
       .find('input[id="id-fieldName"]')
@@ -57,7 +57,9 @@ describe('SelectTextField', function() {
   });
 
   it('can add user input into select field when using paired choices', function() {
-    let wrapper = mount(<SelectTextField choices={[['foo', 'foo']]} name="fieldName" />);
+    let wrapper = mount(
+      <SelectCreatableField choices={[['foo', 'foo']]} name="fieldName" />
+    );
 
     wrapper
       .find('input[id="id-fieldName"]')
@@ -85,7 +87,7 @@ describe('SelectTextField', function() {
     let submitMock = jest.fn();
     let wrapper = mount(
       <Form onSubmit={submitMock}>
-        <SelectTextField choices={[['foo', 'foo']]} name="fieldName" />
+        <SelectCreatableField choices={[['foo', 'foo']]} name="fieldName" />
       </Form>,
       {}
     );
